@@ -1,6 +1,6 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.1.0'
+gem 'rails', '3.2.0'
 gem 'sqlite3-ruby', require: 'sqlite3'
 gem 'devise'
 gem 'dynamic_form'
